fixed .travis.yml

+++ routes/notes.js
@@ -4,6 +4,7 @@ const express = require('express');
 // Create an router instance (aka "mini-app")
 const router = express.Router();
 const Note = require('../models/note')
+const mongoose = require('mongoose')
 /* ========== GET/READ ALL ITEM ========== */
 router.get('/notes', (req, res, next) => {
   const {searchTerm} = req.query
@@ -11,7 +12,6 @@ router.get('/notes', (req, res, next) => {
   filter.search = searchTerm
     ?{$text: {$search: searchTerm}}
     :{}
-  console.log(filter)
   Note.find(
     filter.search,
     {score:{$meta:'textScore'}}
@@ -25,7 +25,13 @@ router.get('/notes', (req, res, next) => {
 });
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/notes/:id', (req, res, next) => {
-  Note.findById(req.params.id)
+  const noteId = req.params.id
+  if(!mongoose.Types.ObjectId.isValid(noteId)){
+    const err = new Error('The `id` is not valid')
+    err.status = 400
+    return next(err)
+  }
+  Note.findById(noteId)
     .then(result=> {
       return res.status(200).json(result)
     })
@@ -39,7 +45,7 @@ router.post('/notes', (req, res, next) => {
   if(err) return next(err)
   Note.create({title,content})
     .then(result => {
-      return res.status(201).json(result)
+      return res.location(`${req.originalUrl}/${result._doc._id}`).status(201).json(result)
     })
     .catch(next)
 });
@@ -50,6 +56,11 @@ router.put('/notes/:id', (req, res, next) => {
   const {title, content} = req.body
   const err = validTitle(title)
   if(err) return next(err)
+  if(!mongoose.Types.ObjectId.isValid(noteId)){
+    const err2 = new Error('not a valid id')
+    err2.status = 400
+    return next(err2)
+  }
   const updateObj = {title, content}
   Note.findByIdAndUpdate(noteId,updateObj,{new:true})
     .then(result => {
@@ -61,11 +72,16 @@ router.put('/notes/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/notes/:id', (req, res, next) => {
   const noteId = req.params.id
+  if(!mongoose.Types.ObjectId.isValid(noteId)){
+    const err = new Error(`${noteId} does not exist`)
+    err.status = 400
+    return next(err)
+  }
   Note.findByIdAndRemove(noteId)
     .then(()=> {
       res.status(204).end()
     })
-});
+})
 module.exports = router;
 
 //========== Validation methods =========================
